Harden the native iOS just-works path and record the prove miss.
Editor uses TextEditor plus a safe-area word-count bar so the caret can sit above the keyboard. Style sheet puts Size last. Library is a stamp-and-row list. Compose ignores a double tap on a fresh blank. XCTest hammer and a glare list are in the tree. xcodebuild is not on this Linux worker (exit 127); the production bar stays false.

+++ ios/VellumPad/VellumPad/Theme/VellumFonts.swift
@@ -27,11 +27,11 @@ enum VellumFonts {
     }
 
     static func title(_ typeface: Typeface, size: TypeSize) -> Font {
-        page(typeface, size: size.titlePoints, relativeTo: .title)
+        page(typeface, size: CGFloat(size.titlePoints), relativeTo: .title)
     }
 
     static func body(_ typeface: Typeface, size: TypeSize) -> Font {
-        page(typeface, size: size.bodyPoints, relativeTo: .body)
+        page(typeface, size: CGFloat(size.bodyPoints), relativeTo: .body)
     }
 
     static func sample(_ typeface: Typeface) -> Font {
